Moving things around to avoid cross-sub-package referencing
