+++ packages/components/src/entities/index.ts
@@ -17,7 +17,6 @@ import {
     getJobCreationHref,
 } from './utils';
 import { SampleTypeBasePage } from './SampleTypeBasePage';
-import { SampleActionsButton } from './SampleActionsButton';
 import { SampleAliquotsSummary } from './SampleAliquotsSummary';
 import { SamplesAddButton } from './SamplesAddButton';
 import { SampleAssayDetail } from './SampleAssayDetail';
@@ -56,7 +55,7 @@ import { SamplesResolver } from './SamplesResolver';
 import { AssayImportSubMenuItem } from './AssayImportSubMenuItem';
 import { useSampleTypeAppContext } from './SampleTypeAppContext';
 import { SampleTypeDesignPage } from './SampleTypeDesignPage';
-import { AssayResultsForSamplesButton, AssayResultsForSamplesMenuItem } from './AssayResultsForSamplesButton';
+import { AssayResultsForSamplesButton, AssayResultsForSamplesMenuItem } from '../internal/components/entities/AssayResultsForSamplesButton';
 import { AssayResultsForSamplesPage, AssayResultsForSamplesSubNav } from './AssayResultsForSamplesPage';
 import { SampleOverviewPanel } from './SampleOverviewPanel';
 import { SampleDetailContextConsumer, SampleDetailPage } from './SampleDetailPage';
@@ -109,7 +108,6 @@ export {
     PicklistOverview,
     PicklistSubNav,
     RemoveFromPicklistButton,
-    SampleActionsButton,
     SampleAliquotDetailHeader,
     SampleAliquotViewSelector,
     SampleAliquotsSummary,
